internal/uritemplates: Implementation of URI Templates level 1
This is a light implementation of the "Level 1" profile of URI Templates as
described in RFC 6570.

This is here to allow for using the standard URI templates language as
part of service definitions in the service discovery protocol. So far
we've only needed literal URLs resolved relative to the discovery document,
but integrating with protocols that were not designed exclusively for
use with Terraform/OpenTofu tends to require more flexibility.

This commit only introduces the new package as dead code. We'll make more
use of it in future commits.