auth: Reorganize the API a little
Experience using and maintaining the other library this one was based on
suggested that the API design was a little too inflexible, with credential
storage (rather than only reading) having been bolted on somewhat-awkwardly
and with the HostCredentials abstraction being insufficient for TLS client
certificate authentication.

Since we're effectively an entirely new library from the Go toolchain's
perspective, we can make some small breaking changes to the API to separate
concerns a little better and also in particular to make room for TLS
client certificate auth, which deals with one significant additional mode
of authentication that the HostCredentials.PrepareRequest design could
not handle, because the TLS handshake precedes making any HTTP requests.

There are not yet any implementations of HostCredentials in this codebase
that can provide TLS client certificates, but downstream users of this
library can provide their own implementations that can combine both
TLS-level and HTTP-level authentication at once if needed.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>