disco: Tidy up some crufty bits
This library started its life as an internal package within OpenTofu's
predecessor, and was factored out into a separate library without much
regard to what parts of it made sense to be maintained outside of that
codebase.

Since we're starting fresh at a new module address we have some freedom to
make some breaking API changes, and so this shrinks the API and makes it
more suitable for use as a separate library. In particular, this removes
some esoteric version handling used only as part of the legacy "remote"
and "cloud" backends in OpenTofu -- which we can reimplement as unexported
code within the OpenTofu codebase instead -- and makes the discovery client
accept a full HTTP client from its caller instead of just a transport,
so the caller can be responsible for selecting an appropriate client
configuration.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>