auth: Remove HostCredentials.ClientCertificate
A previous commit added this as a potential way to support authentication
at the TLS level rather than at the HTTP level, using client certificates.

However, after some attempt to implement this API it's clear that this is
not the appropriate level of abstraction to handle this: TLS client
certificates are a property of the HTTP client rather than of any specific
target server or any single HTTP request.

Instead then, if callers like OpenTofu want to support TLS client certs
for authentication that should be handled as something separate from the
per-host credentials configuration, configuring the HTTP clients created
by the caller in a cross-cutting way rather than in a host-specific way.

A previous commit changed the disco package API to take a full http.Client
from the caller, so the caller can provide a client with a transport that
knows suitable TLS certificates to use.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>