disco: Don't generate log messages
This code started off as an internal package in OpenTofu's predecessor
where it's pretty common to just write directly to the default logger,
but that's not appropriate behavior for a shared library since its caller
might have different opinions about whether and how to collect logs.

For now we'll just remove the logging. If we want to reintroduce some of
this later then we'll introduce a more general API that allows the caller
to provide their own callbacks for certain interesting events, in which
they can handle logging as they wish.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>