disco: DiscoTrace to allow callers to trace requests
Allowing a context.Context to pass through our functions into the HTTP
client is enough to let a caller log or trace the individual HTTP request,
but a request for a particular hostname can potentially cause multiple
HTTP requests to occur if the first request returns a redirect, and so
callers might wish to be able to group together the HTTP requests
associated with a given discovery request.

DiscoveryStart in particular allows the caller to derive a new context
to represent the scope of a particular discovery request, so that e.g.
when using distributed tracing DiscoveryStart can begin a new tracing
span for the child request traces to belong to and then mark that span
as complete in both DiscoverySuccess and DiscoveryFailure.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>