svcauth: CachingCredentialsStore
From some practical experience using CachingCredentialsSource with an
inner source that is statically known to implement CredentialsStore too,
it's annoying for a caller to need to insert what appears to be a dynamic
type assertion for something that can actually be proven at compile time
by the compiler.

Therefore this helper function just encapsulates that infallible dynamic
type check, so that external callers can rely on the compiler to check
they've provided a suitable inner implementation.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>