disco: Tolerate legacy quirks of "tfe" service type.
From the version control history we inherited from OpenTofu's predecessor
in the main OpenTofu repository we can infer that the "remote" backend
had attempted to introduce the idea of minor versions for services, but
then that decision was reversed when that code was copied as the basis
of the special "cloud" backend.

Since "tfe.v2.1" has historically been accepted as a valid service id for
so far, despite the fact that all existing implementers of the API have
set it to refer to the same URL as just "tfe.v2", we'll add a little
legacy shim to treat "tfe.v2.1" as effectively just an alias for "tfe.v2"
moving forward, as a concession to preserve the current functionality of
the remote backend.

Since "tfe" really just represents the API of the commercial TACOS product
offered alongside OpenTofu's predecessor, we aren't likely to use this
service name for anything new in OpenTofu, and minor protocol versions are
better handled by fine-grain service-specific negotiation of individual
features anyway, we'll allow this only for "tfe" and force all other
service types to strictly follow the major-version-only convention, which
is true for all other existing service types and can be true for all new
ones we might adopt in future too.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>